finalize agent

+++ agent/index.js
@@ -5,29 +5,38 @@ const { host, port, role, password, repository, branch } = require('./config')
 
 const client = new net.Socket()
 
+client.on('data', data => {
+	const commandText = data.toString()
+
+	Promise
+		.try(() => {
+			const command = JSON.parse(commandText)
+			if (repository && command.repository.name != repository
+				|| branch && command.branch != branch) {
+				throw new Error(`illegal command ${commandText}`)
+			}
+			return command
+		})
+		.then(command => execute({ command, onFeedback: data => handleFeedback(command, data) }))
+		.catch(handleFailure)
+
+	function handleFeedback (command, data) {
+		send({ type: 'FEEDBACK', command, data })
+	}
+
+	function handleFailure (error) {
+		log.error(`failed to execute ${commandText} due to ${error.stack}`)
+		try {
+			send({ type: 'FAILURE', command: JSON.parse(commandText) })
+		} catch (sendError) {
+			log.error(`failed to send failure due to ${sendError.stack}`)
+		}
+	}
+})
+
 client.connect(port, host, () => {
 	log.debug(`connected to ${host}:${port}`)
 	send({ type: 'AUTHENTICATE', role, password, repository, branch })
-})
-
-client.on('data', data => {
-	const command = JSON.parse(data.toString())
-	if (repository && command.repository.name != repository
-		|| branch && command.branch != branch) {
-		log.error('illegal command %j', command)
-		send({ type: 'ILLEGAL' })
-		return
-	}
-	execute({ ...command, onFeedback: data => send({ type: 'FEEDBACK', command, data }) })
-})
-
-client.on('error', error => {
-	log.error(`connection failed due to ${error.stack}`)
-	process.kill(process.pid, 'SIGINT')
-})
-
-client.on('close', () => {
-	process.kill(process.pid, 'SIGINT')
 })
 
 function send (message) {
@@ -35,7 +44,16 @@ function send (message) {
 	client.write(JSON.stringify(message))
 }
 
-process.on('SIGINT', () => {
-	log.debug('shutting down')
-	client.destroy()
+process.on('SIGINT', shutdown)
+client.on('close', shutdown)
+client.on('error', error => {
+	log.error(`connection failed due to ${error.stack}`)
+	shutdown()
 })
+
+function shutdown () {
+	log.debug('shutting down')
+	if (!client.destroyed) {
+		client.destroy()
+	}
+}
